✅ fix file-explorer test ✅ move db-explorer test into a dedicated test file
